Update: Created own class for public functions of ExcelIOBase. Update: Changed references of public functions from ExcelIOBase to ExcelIOAPIs. Update: Dev Docs.
